Made some optimizations but removed GPU
For some reason when I try run it on the GPU it seems to be slower than if I tried to run it on the CPU. For this reason I am going to disable GPU for now untill I can make it work better. Otherwise there is optimizations that have been made.